Implemented Livewire MeterReadings component

+++ database/migrations/2025_10_03_080008_create_meter_readings_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('meter_readings', function (Blueprint $table) {
             $table->id();
             $table->foreignId('client_id')->constrained()->onDelete('cascade');
-            $table->decimal('previous_meter', 10, 2)->default(0);
-            $table->decimal('current_meter', 10, 2)->default(0);
+            $table->unsignedInteger('previous_meter')->default(0);
+            $table->unsignedInteger('current_meter')->default(0);
             $table->decimal('amount', 10, 2)->default(0); // base monthly bill
             $table->decimal('remaining_amount', 10, 2)->default(0); // unpaid balance for that month
             $table->decimal('maintenance_cost', 10, 2)->default(0); // monthly maintenance
